Moved metal backend files to their own directory, and fixed namespaces
