+++ src/ShaderGen.Tests/ShaderGeneratorTests.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using ShaderGen.Glsl;
 using ShaderGen.Hlsl;
+using ShaderGen.Metal;
 using Xunit;
 
 namespace ShaderGen.Tests
