+++ src/ShaderGen/Metal/MetalBackend.cs
@@ -1,13 +1,11 @@
-﻿using Microsoft.CodeAnalysis;
-using System;
-using System.Text;
+﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-using System.IO;
 using System.Diagnostics;
+using System.Linq;
+using System.Text;
+using Microsoft.CodeAnalysis;
 
-namespace ShaderGen
+namespace ShaderGen.Metal
 {
     public class MetalBackend : LanguageBackend
     {
