+++ src/ShaderGen/Metal/MetalKnownFunctions.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using ShaderGen.Hlsl;
 
-namespace ShaderGen
+namespace ShaderGen.Metal
 {
     public static class MetalKnownFunctions
     {
