+++ src/ShaderGen/Metal/MetalKnownTypes.cs
@@ -1,7 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 
-namespace ShaderGen
+namespace ShaderGen.Metal
 {
     internal static class MetalKnownTypes
     {
