+++ src/ShaderGen/Metal/MetalSwizzleTranslator.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using System.Text;
 using ShaderGen.Hlsl;
 
-namespace ShaderGen
+namespace ShaderGen.Metal
 {
     internal class MetalSwizzleTranslator : TypeInvocationTranslator
     {
